work items

+++ Katiba55.API/Dtos/Projects/ProjectBriefDto.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public double? ExecutionPercent { get; set; }
+        public decimal? ExecutionPercent { get; set; }
         public DateTime? ExecutionDate { get; set; }
         public string ExecutionStatus { get; set; }
     }
