+++ Katiba55.API/Entities/Project.cs
@@ -15,7 +15,7 @@
         public double? Latitude { get; set; }
         public double? Longitude { get; set; }
         
-        public double? ExecutionPercent { get; set; }
+        public decimal? ExecutionPercent { get; set; }
         public DateTime? ExecutionDate { get; set; }
         public ExecutionStatus ExecutionStatus { get; set; }
